grpc-cv-service/pyproject.toml
[project]
name = "grpc-cv-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "grpcio>=1.74.0",
    "grpcio-tools>=1.74.0",
    "onnxruntime>=1.22.1",
    "ultralytics>=8.3.177",
]
